ページ指定等の型をuint -> int に変更。intの方がアプリ側で使いやすい。

+++ Hohoema/Models.Domain/Niconico/Follow/LoginUser/UserFollowProvider.cs
@@ -76,7 +76,7 @@ namespace Hohoema.Models.Domain.Niconico.Follow.LoginUser
                 throw new InvalidOperationException();
             }
 
-            return _niconicoSession.ToolkitContext.Follow.User.GetFollowUsersAsync((uint)pageSize, lastRes);
+            return _niconicoSession.ToolkitContext.Follow.User.GetFollowUsersAsync(pageSize, lastRes);
         }
 
         Task<bool> IFollowProvider<IUser>.IsFollowingAsync(IUser followable) => IsFollowingAsync(followable.Id);
